refactor: Extract getBBox to map helpers
